chore: extract enahance middleware function
